+++ src/activity.h
@@ -302,6 +302,8 @@ public:
     double ret = std::abs(get<double>(value, 0, arrival));
     if (ret == R_PosInf) ret = -1;
     get_resource(arrival)->set_capacity((int)ret);
+    if (!arrival->is_active())
+      return ENQUEUE;
     return 0;
   }