+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Restore `ostream` after formatting (9ff11f8).
 * Fix arrival cloning to copy attributes over to the clone (#118).
+* Fix self-induced preemption through `set_capacity()` (#125).
 
 # simmer 3.6.4